change import of class that moved
