+++ projects/adapter/src/dbt/adapters/fal/impl.py
@@ -46,7 +46,7 @@ class FalEncAdapter(BaseAdapter):
     # TODO: how do we actually use this?
     AdapterSpecificConfigs = FalConfigs
 
-    def __new__(cls, config):
+    def __new__(cls, config, mp_context):
         # There are two different credentials types which can be passed to FalEncAdapter
         # 1. FalEncCredentials
         # 2. FalCredentialsWrapper
@@ -83,7 +83,7 @@ class FalEncAdapter(BaseAdapter):
         with _release_plugin_lock():
             # Temporary credentials for register
             config.credentials = config.sql_adapter_credentials
-            FACTORY.register_adapter(config)
+            FACTORY.register_adapter(config, mp_context)
             config.credentials = FalCredentialsWrapper(config.sql_adapter_credentials)
 
         return FalEncAdapterWrapper(db_adapter_class, config)
