+++ projects/adapter/src/dbt/adapters/fal_experimental/connections.py
@@ -2,8 +2,8 @@ from dataclasses import dataclass
 from typing import Optional
 import os
 
-from dbt.adapters.base import Credentials
-from dbt.dataclass_schema import StrEnum, ExtensibleDbtClassMixin
+from dbt.adapters.contracts.connection import Credentials
+from dbt_common.dataclass_schema import StrEnum, ExtensibleDbtClassMixin
 
 from dbt.fal.adapters.python import PythonConnectionManager
