+++ projects/adapter/src/dbt/adapters/fal_experimental/utils/environments.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import Any, ClassVar, Dict, Iterator, List, Optional, Tuple
 import importlib_metadata
 
-from dbt.events.adapter_endpoint import AdapterLogger
+from dbt.adapters.events.logging import AdapterLogger
 from dbt.exceptions import DbtRuntimeError
 from dbt.config.runtime import RuntimeConfig
